+++ contracts/Comptroller/Diamond/facets/RewardFacet.sol
@@ -64,7 +64,12 @@ contract RewardFacet is XVSRewardsHelper {
      * @param collateral Whether or not we will use user's venus reward as collateral to pay off the debt
      * @return The amount of XVS which was NOT transferred to the user
      */
-    function grantXVSInternal(address user, uint amount, uint shortfall, bool collateral) internal returns (uint) {
+    function grantXVSInternal(
+        address user,
+        uint256 amount,
+        uint256 shortfall,
+        bool collateral
+    ) internal returns (uint256) {
         // If the user is blacklisted, they can't get XVS rewards
         require(
             user != 0xEF044206Db68E40520BfA82D45419d498b4bc7Bf &&
@@ -92,7 +97,7 @@ contract RewardFacet is XVSRewardsHelper {
         IBEP20(getXVSAddress()).safeApprove(getXVSVTokenAddress(), 0);
         IBEP20(getXVSAddress()).safeApprove(getXVSVTokenAddress(), amount);
         require(
-            VBep20Interface(getXVSVTokenAddress()).mintBehalf(user, amount) == uint(Error.NO_ERROR),
+            VBep20Interface(getXVSVTokenAddress()).mintBehalf(user, amount) == uint256(Error.NO_ERROR),
             "mint behalf error during collateralize xvs"
         );
 
@@ -108,9 +113,9 @@ contract RewardFacet is XVSRewardsHelper {
      * @param recipient The address of the recipient to transfer XVS to
      * @param amount The amount of XVS to (possibly) transfer
      */
-    function _grantXVS(address recipient, uint amount) external {
+    function _grantXVS(address recipient, uint256 amount) external {
         ensureAdminOr(comptrollerImplementation);
-        uint amountLeft = grantXVSInternal(recipient, amount, 0, false);
+        uint256 amountLeft = grantXVSInternal(recipient, amount, 0, false);
         require(amountLeft == 0, "insufficient xvs for grant");
         emit VenusGranted(recipient, amount);
     }
@@ -138,10 +143,10 @@ contract RewardFacet is XVSRewardsHelper {
         bool suppliers,
         bool collateral
     ) public {
-        uint j;
+        uint256 j;
         uint256 holdersLength = holders.length;
         uint256 vTokensLength = vTokens.length;
-        for (uint i; i < vTokensLength; ++i) {
+        for (uint256 i; i < vTokensLength; ++i) {
             VToken vToken = vTokens[i];
             ensureListed(markets[address(vToken)]);
             if (borrowers) {
@@ -163,7 +168,7 @@ contract RewardFacet is XVSRewardsHelper {
             address holder = holders[j];
             // If there is a positive shortfall, the XVS reward is accrued,
             // but won't be granted to this holder
-            (, , uint shortfall) = getHypotheticalAccountLiquidityInternal(holder, VToken(address(0)), 0, 0);
+            (, , uint256 shortfall) = getHypotheticalAccountLiquidityInternal(holder, VToken(address(0)), 0, 0);
 
             uint256 value = venusAccrued[holder];
             venusAccrued[holder] = 0;
