+++ contracts/Comptroller/Diamond/interfaces/IRewardFacet.sol
@@ -14,9 +14,9 @@ interface IRewardFacet {
 
     function claimVenusAsCollateral(address holder) external;
 
-    function _grantXVS(address recipient, uint amount) external;
+    function _grantXVS(address recipient, uint256 amount) external;
 
-    function getBlockNumber() external view returns (uint);
+    function getBlockNumber() external view returns (uint256);
 
     function getXVSAddress() external pure returns (address);
